+++ plugin-name/admin/class-plugin-name-admin.php
@@ -70,27 +70,29 @@ class Plugin_Name_Admin {
 		$plugin = Plugin_Name::get_instance();
 		$this->plugin_slug = $plugin->get_plugin_slug();
 		$this->version = $plugin->get_plugin_version();
-		
-		//Search update
+
+		//Check update of plugin
 		require_once( plugin_dir_path( __FILE__ ) . '/includes/TGM-Updater/updater/init.php' );
 		$args = array(
-			'plugin_name' => 'Soliloquy', // Your plugin name (e.g. "Soliloquy" or "Jetpack")
+			'plugin_name' => $plugin->get_plugin_name(), // Your plugin name (e.g. "Soliloquy" or "Jetpack")
 			'plugin_slug' => $this->plugin_slug, // Your plugin slug (typically the plugin folder name, e.g. "soliloquy")
 			'plugin_path' => plugin_basename( __FILE__ ), // The plugin basename (e.g. plugin_basename( __FILE__ ))
-			'plugin_url' => WP_PLUGIN_URL . '/plugin-name', // The HTTP URL to the plugin (e.g. WP_PLUGIN_URL . '/soliloquy')
+			'plugin_url' => WP_PLUGIN_URL . $this->plugin_slug, // The HTTP URL to the plugin (e.g. WP_PLUGIN_URL . '/soliloquy')
 			'version' => $this->version, // The current version of your plugin
-			'remote_url' => 'http://soliloquywp.com/', // The remote API URL that should be pinged when retrieving plugin update info
-			'time' => 42300						 // The amount of time between update checks (defaults to 12 hours)
+			'remote_url' => 'plugin-domain-url.dev', // The remote API URL that should be pinged when retrieving plugin update info
+			'time' => 42300 // The amount of time between update checks (defaults to 12 hours)
 		);
 		$config = new TGM_Updater_Config( $args );
 		$namespace_updater = new TGM_Updater( $config ); // Be sure to replace "namespace" with your own custom namespace
-		$namespace_updater->update_plugins();			// Be sure to replace "namespace" with your own custom namespace
+		$namespace_updater->update_plugins();   // Be sure to replace "namespace" with your own custom namespace
 		// Load admin style sheet and JavaScript.
 		add_action( 'admin_enqueue_scripts', array( $this, 'enqueue_admin_styles' ) );
 		add_action( 'admin_enqueue_scripts', array( $this, 'enqueue_admin_scripts' ) );
 
 		// Add the options page and menu item.
 		add_action( 'admin_menu', array( $this, 'add_plugin_admin_menu' ) );
+		//Add the plugin settings link in the plugins list page
+		add_filter( 'plugin_action_links_' . $this->plugin_slug . '.php', 'add_plugin_settings_link' );
 
 		// Add an action link pointing to the options page.
 		$plugin_basename = plugin_basename( plugin_dir_path( realpath( dirname( __FILE__ ) ) ) . $this->plugin_slug . '.php' );
@@ -225,6 +227,12 @@ class Plugin_Name_Admin {
 		$this->plugin_screen_hook_suffix = add_options_page(
 				__( 'Page Title', $this->plugin_slug ), __( 'Menu Text', $this->plugin_slug ), 'manage_options', $this->plugin_slug, array( $this, 'display_plugin_admin_page' )
 		);
+	}
+
+	function add_plugin_settings_link( $links ) {
+		$settings_link = '<a href="options-general.php?page='.$this->plugin_slug.'">' . __( 'Settings', $this->plugin_slug ) . '</a>';
+		array_push( $links, $settings_link );
+		return $links;
 	}
 
 	/**